Adds the ability to track dirty diapers.
There are a few hidden config files here, since I don't want to
publicly share my password / details of my baby.

This has been tested locally as a Python application, but I
haven't uploaded this version to AWS to confirm that it works
as a lambda function (much less and Alexa app).